Adds logo to SideBar
Adds a logo to the SideBar component. Udpates the SideBar component to
have more room in the top and bottom flares. Adds a small version of the
logo to the Logo component to allow for a more specific fit.

Creates the scroll utils module file to make the logic for showing
components when not in the landing section reusable.