Updates 404 page styling
Updates the styling of the 404 page by removing the default text and
centering the h1 element.

Updates the Header and SideBar to only hide the header and logo in the
SideBar at the top of a page if that page is the Home page (or another
specified page, currently no others). This is done so that the header
and logo in the SideBar appear on the 404 page by default.

Creates the LocationContext to accomplish this. The LocationContext
stores information about the page location.