Updates references to window for build
Updates references to the `window` object so that they are not used at
build-time (component instantiation), only at run-time.

Adds the `useMounted` hook, which packages the logic to set a state
variable (`isMounted`) to `true` when a component mounts. Makes use of
this hook for updating window-based variables after mount.